Consolidate various utility files into a single helper.py - Refactoring imports (isort) - Removing duplicate code
