+++ wikimedia/entries/entry.py
@@ -1,6 +1,6 @@
-
 from trackers.tracker import Tracker
-from utilities.fs import FileSystem
+from utilities.helpers import ParquetHelper
+
 
 class Entry():
     """
@@ -11,7 +11,7 @@ class Entry():
         """
         Load data from parquet file and filter out ids if a file filter is provided
         """
-        fs = FileSystem()
+        fs = ParquetHelper()
         data = fs.read_parquet(data_in, cols=columns)
 
         if file_filter:
